Fix all file headers
All the "rulers" in the file headers end now at column 78 and their
length does not depend on the length of the single-line comment starter.

The init_file tool becomes simpler (one less parameter and simpler
logic).

Incidentally, a minor issue in the documentation of prefixed_dump.awk
(repetition of the word "integer") is fixed and missing boilerplate in
find_environment_string_test.cpp is added.