Use the WG21 redirect service anywhere possible (for links in comments)
Reason: keeping the URLs within 80 columns.

Note that for is_power_of_2() we were also being inconsistent, by using
the redirect service in one case and not in the other one; and that we
still have a link to open-std.org in the codebase, but that's for C, for
which there's no redirect service (AFAIK).