Don't change the status of a moved from maybe
Reason: in case the other object has a value, we move from that value,
but that doesn't make the maybe object "value-less", so we cannot set
the status to invalid; in case other has no value, why risking changing
the "reason for failure"?

Note that, for the move constructor, we had to fix the documentation;
for the move assignment operator, this was not necessary.