Don't delay setting the status in maybe's constructors
Reason: delaying allowed having asserts in the two construct()
overloads, to check that we didn't construct a new object "over" an
existing one; OTOH, it might make the code less clear. Are the two
asserts excessive precaution? (I'm not sure; I might in fact revert this
commit).

Note that a comment has been moved from the move constructor to the copy
constructor, because the copy constructor is defined first.