Don't give a possibly misleading message when FormatMessage() fails
When FormatMessage() fails there are two errors involved: the original
one and the FormatMessage() failure.

The previous message could lead the user to think that the error code is
for the FormatMessage() failure. It is still the one for the original
error, instead.

Note that the old message missed "()" after "FormatMessage", but since
we have changed it completely...

The next commit will sync this with the Unix implementation.