Nest all unnamed namespaces in namespace breath (only in library code)
Reason: consistency (we were doing it in some places and not in others).
I'm not sure this will make a practical difference when breath will be
an alias for a "versioned name", such as breath_2020v1, etc.

Note that this doesn't affect test files, examples, etc.: it only
applies to library code.