Remove a facility: aligned_storage_for
Reason: it no longer makes sense; it was used by the maybe template when
we used C++98/03, but maybe is using several C++11 features now, so why
not using alignas, too. And aligned_storage_for itself is using noexcept
and BREATH_DECLARE_NON_COPYABLE(), so it wouldn't compile as C++98/03
(this occurred probably while evolving the code, and was an oversight).