Include the meta::integer_log2 tests in the automatic test run
Reason: this is another "piece" we missed in the previous commits.

Note that I didn't bother renaming integer_log2_compile_time_test.cpp,
or adding a note that the test is actually compile-time but we made it
run-time to have a report, because I'm going to drop meta::integer_log2
and support for MSVC 2015.