Report OS and compiler version for the Travis CI FreeBSD build job
I needed to see the Travis CI output to know these, so I didn't add them
with the previous commit. In fact, the possibility to use FreeBSD
doesn't seem to be documented, and it doesn't work (I saw the
.travis.yml of another project apparently using FreeBSD and I decided to
give it a try). I might remove the build job, now. Or wait for a while,
to see if Travis fixes its support.

Other keywords: description, name.