Remove a facility: BREATH_DECLARE_NON_COPYABLE()
Reason: although I sort of liked the expressiveness of the macro, some
programmers expressed dislike, on the ground that it "hides" the
established way to disable copying and forces them to learn a
library-specific idiom to understand the library.

Note that Doxygen had problems with the macro, but it has them without
the macro, too (I can see the copy constructor among the members but not
the copy assignment operator).

Other keywords: = delete.