Group sha_common.tpp with the other .tpp files in Visual Studio
Reason: making it easier to find for the users who search "visually"
through Solution Explorer.

The file was in "Source Files", but it was the only .tpp to be there
(after this commit we'll only have .cpp's there; which I'm not sure is
right, but at least it's consistent).

Other keywords: project filters.