Add many \brief Doxygen descriptions
Reason: the "Class and Class Templates" list, in the generated
documentation, would show almost nothing, otherwise.

Note that we often copybrief'd the file-level description, and that the
only descriptions we did not add are those for the test and the command
line facilities, which are not documented, anyway. Also, we hid some
facilities as implementation details (e.g. has_sign< bool >), to remove
them from the list page mentioned above.

Finally, note that we had to "inform" Clang about the \copybrief
command.