Don't use "main.cpp" as file name for our tools
Reason: while I like "main.cpp" as a file name (the containing directory
name provides the context: soo foo/main.cpp is the main file of the
"foo" tool), it means that, with our current build system, we could only
build one tool (we can't have multiple source files with the same name).
So, I'm changing convention: the main file of the foo tool is foo.cpp.
This is also more consistent (though not fully consistent) with the
choice to name the main unit test file "unit_test_main.cpp".