Make $(dependency_dir) an order-only prerequisite
Reason: it avoids that the directory is created as soon as the makefile
is parsed, thus also where no file is actually built (e.g. in
BREATH_ROOT).

Apparently, the linked to article on mad-scientist.net was updated in
the same vein, because it's also using a read-only prerequisite, now
(and I think it was creating the directory on parsing, when I originally
looked at it; in fact, it currently reports that it was updated on July
4th, 2020---too bad that I didn't save the version I used initially).

Other keywords: build system, GNU Make.