Relax the requirements on iterators used with endian_codec
To enable using non-random-access iterators, we changed approach.

The previous approach required the endianness policy to produce the
indexes at which to write or read the "pieces" (i.e. the Bytes) of the
value.

The new one requires producing the shift amounts, instead. This allows
reading and writing always to/from *iter, then *++iter, etc., in order.
So we no longer need a random access iterator.

Of course, we also add tests using input and output iterators.

PS: there's still an index() member in both policies, which is used by
the cryptography parts; I'll see if I can rewrite those parts to use a
similar approach.