Don't restrict is_power_of_2() to unsigned types
Reason: the SFINAE was awful, and I don't know if the restriction was
even opportune (it was done to mimic the standard, but does the standard
get this right?).

The tests now use the type int in lieu of unsigned int. The next commit
will add tests using other types.