Fix the const correctness logic of auto_array
This makes const apply exactly where the user applies it: i.e. to the
auto_array (i.e. to the managed pointer) when the auto_array is const,
or to the pointed to elements when T is const. Previously, you couldn't
assign to the elements of a non-const array through a const auto_array.