Change last_api_error to use the string managed by the base class
As announced in the commit "Redesign all our usage of exceptions", this
removes the member m_message from last_api_error, and uses the string
managed by the base class, instead. Note that, as a consequence, we no
longer need to override what(). Also, since we got rid of C-style
strings, we just use std::string in the new format_message() function.

See the code comments for why we added a private constructor.