Rewrite nist_vectors_test.cpp
Reason: the main loop was a mess.

The new implementation uses two separate classes for the message and the
hash files, both derived from nist_file, which just expose a next()
function and encapsulate the handling of "D>" (start of a section) and
"<D" (end of a section). This largely simplifies the main loop.

Other keywords: unit tests.