Fix the markup of all the Doxygen \throw commands
The \throw command has a mandatory argument, which is the type of
exceptions emitted. When we followed "\throw" with something like "A \c
std::runtime_error...", then "A" was taken as the argument, and
formatted as code. See <https://github.com/erez-o/docsforge/issues/34>
for a screenshot.

Note that one of the paragraphs was transformed into a note, as we had
no exception type to indicate.