Fix all the \throw paragraphs in the documentation
The \throw command has a mandatory argument, which is the type of
exceptions emitted. When we followed "\throw" with something like "A \c
std::invalid_argument", then "A" was taken as the argument, and
formatted as code.

Note that we transform one of the paragraphs into a note, as we had no
exception type to indicate.

Other keywords: Doxygen.