Allow each GNUmakefile to specify additional libraries to link with
The immediate need for this is to specify Version.lib on Windows (in the
next commit). So far, we have only needed to link with Advapi32.lib, and
since Clang and GCC seemed to link that automatically (at least on
Cygwin) we specified Advapi32.lib in test/GNUmakefile and
tool/include_guard/GNUmakefile and handled that for MSVC only (which, of
course, implies Windows). Now we need better.

The solution we adopt for now is that any GNUmakefile which needs to
specify additional libraries to link with defines the variable
'libraries' to a list of their names (without prefixes and
suffixes---see the new comments in main.mk)

As to Advapi32.lib specifically, note that, for simplicity, we
explicitly link it even with Clang and GCC (only on Windows, of course),
although that doesn't appear to be necessary, at least with Cygwin.