Make last_api_error copy assignable
Reason: I had deleted the copy assignment operator on the ground that
nobody assigns exception objects; however, last_api_error can be used
also as a "normal" (i.e. non-exception) object and, more importantly,
deleting the copy assignment operator makes the generation of the
implicit copy constructor deprecated, so we had to declare that
(possibly as = default), but we want to get rid of noise in the class
declaration, and we'll do so with the next commit.