Make gcd() compile without warnings on Clang and GCC
Clang and GCC gave -Wsign-compare warnings when the arguments to gcd()
had no sign and we compared them to -1 (although we first checked if
they were negative). So we introduced a SFINAE'd utility that avoids the
comparisons in the no-sign case. But this produced a regression with
MSVC 2017, so we added a workaround for that, too.