Remove or reduce our file trailers
Reason: using EditorConfig only is simpler, easier to maintain, and
removes noise from the files.

This removes all file trailers, except, as explained in the comments to
init_file.sh, for leaving the pieces that set the file type when the
extension we use isn't recognized by Vim or Emacs (this happens with
".tpp", ".shc" and ".dox").

Note that, for CSS, HTML, XHTML and XSLT, and for Markdown (not yet
supported by the init_file tool), the file trailers were also used to
set UTF-8 (with no BOM), but this seems unnecessary, so we don't do it
via EditorConfig, either; and, among those types, we are only using HTML
and Markdown. And that we also removed the trailer from doxygen.cfg.