Start migrating from Travis CI to GitHub Actions
Reason: Travis CI is no longer free (and doesn't work well, anyway).

Note that this is not a "translation" of the Travis CI configuration (we
had no reason to keep exactly that one, and we wanted to start with
something simpler anyway, because it's difficult to get everything right
from start).

With respect to the Travis CI configuration, we build with less
compilers, for the moment, and without sanitizers, but build on macOS,
too.