Fix the Doxygen documentation for some "grouped members"
The fixes consist either in separating the \name line, so that Doxygen
understands that what follows is the documentation of the first member
(as done in crc.hpp), or in adding missing documentation for a member
(as done in merkle_damgard_machine.hpp) or in adding \name commands.

Other keywords: markup.