Use _SC_PAGESIZE in lieu of _SC_PAGE_SIZE
Reason: the FreeBSD headers don't seem to define the latter, at least by
default; I hope they define the former (that's what I've been led to
believe by googling the issue a bit). There might be a way to get the
latter (which reads better, but is otherwise a synonym), but I didn't
find anything online in this regard, and I don't have access to a
FreeBSD machine other than through GitHub Actions.