Define _XOPEN_SOURCE to 700 for Clang and GCC builds
Reason: this should cause _POSIX_C_SOURCE to be defined to 200809L,
which in turn should cause __DARWIN_C_LEVEL to be defined to the same
value, thus making the definitions of

  ENOTRECOVERABLE

and

  EOWNERDEAD

available in errno.h, on macOS, without which our builds fail.

Note that this commit effectively subsumes "Enable the XSI extensions
for GCC and Clang builds".