Distinguish between 'system' and 'system_family' for file selection
Reason: so far, it has been enough to distinguish between Windows and
Unix. Now, we have a facility that doesn't work on macOS with
system=unix (amount_of_physical_memory(), as the macOS headers don't
define _SC_PHYS_PAGES). So, we need a more fine-grained distinction.

Note that we avoided renaming all the syst/ directories to something
including "family". We might do that in the future.

Other keywords: variants, build system, dependent sources.