Make the cryptographic function #include's easier to use
Reason: the previous organization was a bit confusing, in that the
*_hasher.hpp files defined the hashing policy and a typedef for the
hasher but also a typedef for the digest type (which one wouldn't expect
from the file name); furthermore, the latter could be a typedef to an
incomplete type. So, if e.g. you wanted to define a digest object, you
still needed to include digest.hpp. Now <function-name>.hpp provides
everything. This is a violation to our usual "one thing per file"
policy, but I think it's worth doing.

For consistency, this also renames the corresponding .cpp files.