Give the FreeBSD virtual machine 2 GiB of RAM, on GitHub Actions
Reason: one of our SHA-384 tests uses a 1 GiB string.

Note that we had introduced a check for the amount of RAM with

  "Skip one of the SHA-384 tests if the machine doesn't have enough
  RAM".

This commit removes that check, as it is no longer necessary. The
possibility to increase the amount of RAM has been introduced with the
new version of vmactions/freebsd-vm, at my request.

Other keywords: continuous integration.