Redesign os_id as a scoped enum
Reason: the static members of os_id were not safe with regard to order
of initialization issues. This could have been solved by making them
constexpr and non-members (because, apparently there is no way to make
them constexpr and members---maybe a defect in the standard?) but, then,
we would have needed inline constexpr, which isn't in C++14; anyway, why
going to that route if an enum class just works?