Remove silly usages of 'std::string::size_type'
Reason: std::string uses std::allocator, so size_type is guaranteed to
be the same as std::size_t. The forms we were using could make sense if
there was a desire to make the code ready for generalizing it to use
basic_string, but now we know that we'll never do that generalization.