Make our stream inserters handle width() and fill() correctly
We had an issue with inserters that internally did multiple insertions.
Specifically, the inserters of:

 - last_api_error

 - operating_system_name for Unix

 - operating_system_id

 - printable_string

They would apply width() and fill() to the first of those insertions,
only, which is incorrect.

A fix for this is to output to an ostringstream, let's call it "oss",
first, then do dest << oss.str(). That means, though, that our general
to_string() template would again output to an ostringstream and take
str(). IOWs, a conversion to string would imply twice the same
operations (an output to an ostringstream, then str(), then an output to
an ostringstream, then a str()).

So, we devised a simple solution: these types would offer a to_string()
member, and breeze::to_string() would use such a member, if available;
in addition, operating_system_id, which can't have member functions
because it is an enum, would overload breeze::to_string(). The inserters
would simply be implemented in terms of to_string().

Note that class roman, instead, handles width() and fill() correctly,
because it builds a string, first. But it's worth to give it a
to_string() member, as well, to return that string directly. This will
be done in the next commit.

Finally, note that some error checking is missing, because I'd like to
avoid duplicating it in several places, so I want to think of a good way
to do that.