Support arrays and containers of 2**31 elements or more in 32-bit builds
Reason: although nobody will probably have such an array, it can be
declared and used. Note that we are not supporting 16-bit machines very
well by doing this, because using 64 bit ints on such a machine is
likely to be very slow, but we wouldn't support them well by using
std::ptrdiff_t, either. IOWs, they are out anyway. So, let's use
something that works well on 32-bit and 64-bit machines.