Improve and streamline all our Doxygen markup
This commit introduces Markdown for code spans and snippets, and
replaces some uses of "<code>" and similar with "<tt>" (e.g. for
interval designations).

Reason: Markdown is more readable, especially when a code span is
composed of multiple words; interval designations, etc. are not "code".

This commit isn't as "clean" as usual:

 - I re-wrapped the comments after the various replacements, so I have,
   in fact, also re-wrapped comments that I hadn't touched :-s

 - there may be other minor tweaks I did along the way: this patch was
   extremely boring, so I got a bit sloppy