Clarify an error message of our unit tests
The message reports the number of "root" test functions that failed.
Actually each "root" function calls other test functions, so the number
of failing functions can be more than reported. However, the number of
"root" test functions that failed is equal to the number of components
that have failing tests.