Make the endian policies have just one member
This basically reverts "Relax the requirements on iterators used with
endian_codec".

Reason: conceptually, a policy just needs to provide either the shift
amounts to use for encoding/decoding, or the indexes at which to
write/read. Due to some code in merkle_damgard_engine::encode_length,
though, we had two members. We could have replaced usage of index() in
the cryptography parts by something else, but I was too bored to think
of a way to achieve that. In the end, allowing more general iterators
was probably an over-generalization, so why bothering. This basically
brings endian_codec back to the original design, except that each policy
now only has one member function template.