Remove a facility: secure_fill()
Reason: it doesn't work in all cases :-(.

I tried the following on godbolt.org, with GCC 10.2, x86_64:

    template< typename T >
    T volatile &
    volatilize( T & t )
    {
        return t ;
    }

    template< typename Iter >
    void
    my_erase( Iter begin, Iter end )
    {
        while ( begin != end ) {
            volatilize( *begin ) = 0 ;
            ++ begin ;
        }
    }

    int
    main()
    {
        char x[] = "test" ;

        my_erase( x, x + sizeof x ) ;
    }

and the assembly output clearly showed that the five elements of `x`
were zeroed out. But changing "test" to "testing" was enough to have the
compiler remove any write to the array. So, regardless of what the
standard says, our technique doesn't work in the real world.

Note that this means we also remove `sensitive_buffer`, which was a
private component, and that we no longer attempt to zeroize sensitive
information in the cryptographic facilities. I might have attempted to
use `explicit_bzero()` and `SecureZeroMemory()`, but, in all honesty, it
seems that the problem of erasing sensitive information is much more
complicated than expected, even if you erase it in some places, so I
give up.