Use a more precise language in the 'Generalities' section of the docs
Basically, all library facilities except macros are in namespace
`breeze`. The old wording said "most" because we had a global function
(`get_cpuid_info()`). We no longer have it, but saying that all the
facilities are in `breeze` would not be correct, either, because macros
aren't in any namespace. Hence the new wording.

Other keywords: documentation.