When possible, avoid warnings by using #pragma's
Reason: clearer.

So far, we usually avoided warnings either by disabling them globally,
or massaging the code. When using a #pragma is clearly better, we now do
that. (We still disable the warning or massage the code when too many
compilers give warnings about the same piece of code. I tried writing
two general macros for pushing and popping the warning state, so that
one could write e.g.

  # pragma BREEZE_PUSH_DIAGNOSTIC
  <... compiler specific pragmas...>
  # pragma BREEZE_POP_DIAGNOSTIC

instead of having a pair of push/pop for each compiler; but that's not
possible, apparently, not even by using _Pragma().)