Renounce making our exit codes extern objects
Reason: the original motivation for making them extern was to avoid
mismatches if the values were ever changed and an application using the
new values was linked against a version of the library using the old
values. Making them extern, however, opens the door to
order-of-initialization problems. And, indeed, we have a global array in
program.cpp which contains the exit codes, which could, in fact, be
initialized before the exit codes themselves (because those are in
another translation unit). So, we give the codes internal linkage,
relying on the fact that we encode the library version in the namespace
name to prevent linking mismatches.

Note that we have tweaked the Doxygen configuration to have it show the
constants in the documentation, however this only has effect on
exit_success; for the other constants, I don't know what to do yet: the
files they are defined in are in the */dep/* directories, which are
excluded, but, if we include them, we get duplicated entries (because
they are taken from both the Windows and the Unix variant).