Redesign format_time()
Reason: the new version works in terms of time_points, not time_t's;
also, it has no size limitations and takes into account locales.

In fact, adding the time_t parameter to the function was an error: where
are you supposed to get that value for anything else than the *current*
time? From mktime()? But, then, you'd be constructing a time_t from a
tm, which would then be re-converted to a tm by the function. So, it
seems that either you have the current time or a tm; and thus that the
additional argument to format_time() should be a tm.

But since, starting from C++11, we are supposed to work with
time_point's, I redesigned the function in a more "modern" way.

The new version does without strftime() and uses std::put_time(),
writing to a std::ostream, rather than generating a string. I'd have
preferred the latter, because that's handier, for instance, if you want
to put the formatted time in a GUI label. But even using std::time_put()
directly (as opposed to put_time()) requires a stream, so...

There are two solutions to this:

 - using strftime() again (which means giving up considering
   std::locale's)

 - adding an overload of breeze::to_string() taking a
   system_clock::time_point and using format_time()