Remove name() and set_name() from class program
Reason: this solution is much simpler: the user just provides a function
that returns the program name and class program calls it when needed
(which is when outputting an error message).

Note that the user will not usually need to get the name from argv[ 0 ]:
they will just use a fixed name, which they will know when building the
program.